Split up Project module
